Update #19 , #26

+++ codigo/App.java
@@ -327,7 +327,7 @@ public class App{
     // Ainda a testar !
     public static double kmMediaTodasRotas(Frota frota){
         
-        int media =  frota.getVeiculos().stream()
+        double media =  frota.getVeiculos().stream()
         .mapToDouble(Veiculo :: quilometragem)
         .average()
         .getAsDouble();
